backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
alembic
python-jose[cryptography]
passlib[bcrypt]
httpx
google-generativeai
chromadb
python-multipart
langchain
langchain-community
langchain-google-genai
pypdf
tiktoken
sentence-transformers
